+++ examples/demo_revcomp.rs
@@ -4,13 +4,13 @@ use zoe::prelude::*;
 fn main() {
     let args: Vec<String> = env::args().collect();
 
-    let filename = if args.len() == 2 {
+    let path = if args.len() == 2 {
         args[1].clone()
     } else {
         "examples/example.fasta".to_owned()
     };
 
-    let fasta_reader = FastaReader::from_filename(filename).unwrap_or_die("RevComp file error!");
+    let fasta_reader = FastaReader::from_path(path).unwrap_or_fail();
 
     for record in fasta_reader {
         let mut record = record.unwrap_or_die("RevComp FastaReader error.");