+++ src/data/records/fasta/mod.rs
@@ -316,18 +316,34 @@ impl<R: std::io::Read> FastaReader<R> {
 }
 
 impl FastaReader<std::fs::File> {
-    /// Reads a fasta file into an iterator backed by a buffered reader.
+    /// Deprecated, use [`from_path`] instead.
+    ///
+    /// [`from_path`]: FastaReader::from_path
+    #[allow(clippy::missing_errors_doc)]
+    #[deprecated(
+        since = "0.0.27",
+        note = "please use `from_path` instead. This function will be removed in v0.0.29"
+    )]
+    pub fn from_filename<P>(path: P) -> std::io::Result<FastaReader<File>>
+    where
+        P: AsRef<Path>, {
+        Self::from_path(path)
+    }
+
+    /// Creates an iterator over the FASTA data contained in a path, using a
+    /// buffered reader.
     ///
     /// ## Errors
     ///
-    /// Will return `Err` if file or permissions do not exist, or if the file is
-    /// empty. The file path is included in the error message.
-    pub fn from_filename<P>(filename: P) -> std::io::Result<FastaReader<File>>
+    /// Will return `Err` if the path does not exist, if there are insufficient
+    /// permissions to read from it, or if it contains no data. The path is
+    /// included in the error message.
+    pub fn from_path<P>(path: P) -> std::io::Result<FastaReader<File>>
     where
         P: AsRef<Path>, {
-        let path = filename.as_ref();
-        let file = File::open(path).with_file_context("Failed to open path", path)?;
-        Ok(Self::from_readable(file).with_file_context("Failed to read data at path", path)?)
+        let path = path.as_ref();
+        let file = File::open(path).with_path_context("Failed to open path", path)?;
+        Ok(Self::from_readable(file).with_path_context("Failed to read data at path", path)?)
     }
 }